Okay we guuud
Today i was working on figuring out what is happening with the movement
validation function. I was still in the testing stages, and we reached
the problem where my test char (in the lower corner) would not get
detected correctly by the function. I was going to test if it was just
the mouse and the char not being in the correct positions